Fixed statcoulomb definition, which was expressed in units of SI equivalence, but not in units that are meaningful to cgs.
Make angles regular unit quantities with dimensionless reference values.
It doesnt really make sense to make them irreducible quantities.

Added a _reference property, masked because it is really only intended to
be used behind the scenes to calculate conversion factors. The _reference
property was necessary in order to implement customizable default units.

Added a set_default_units function, which allows the user to customize 
how the units of a simplified quantity are expressed, for example in SI
or cgs, or some other customized system.