Remove deprecated mb as symbol for millibar unit
In order to be consistent with the other pressure units that are derived
from the unit bar, we should use mbar as the symbol for millibar. The
symbol mb, which is a deprecated symbol, because the b in mb conflicts
with the metric unit barn, was removed. Since the other bar units also
define an alias with a written out SI prefix, this change also adds the
unit "millibars" as an alias for mbar.