Improved behavior of comparison between quantities
For quantities A and B of unequal dimensionality, the `==` operator
now returns an array of the same shape that numpy would return
if A and B were different and filled with False.

In particular, if both A and B are scalar arrays, the result is
a scalar False. Previously, a scalar array (`np.array(False)`)
would be returned under these circumstances.

If A and B do not have compatible sizes, an exception is generated.

The `!=` operator behaves the same except that it fills the result
with True values.

The other comparison operators now raise exceptions if a quantity
that is not dimensionless is compared to a plain number. For
effectively dimensionless quantities the comparison is now
consistent. Previously, if
    n = pq.s * pq.kHz,
then n < 2 would return True but n.simplified < 2 would return
False. Now, both return False.

Comparison between incompatible quantities raises exceptions as
before.